# docs/03-api/openapi.yaml
openapi: 3.1.0
info:
  title: Git Migration Support Platform API
  version: 0.1.0
  summary: Internal API for GitHub Enterprise migration support and quality operations
  description: >-
    Internal API for repository state visualization, PR risk analysis,
    conflict assistance, backout workflows, policy management, and dashboards.
servers:
  - url: https://git-support.internal.example.com
    description: Production
  - url: https://git-support.staging.internal.example.com
    description: Staging
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/pull-requests:
    get:
      tags: [PullRequests]
      summary: List pull requests with filters
      operationId: listPullRequests
      parameters:
        - in: query
          name: repositoryId
          schema:
            type: string
            format: uuid
        - in: query
          name: state
          schema:
            type: string
            enum: [open, closed, merged, draft, all]
        - in: query
          name: assignee
          schema:
            type: string
        - in: query
          name: riskLevel
          schema:
            $ref: '#/components/schemas/RiskLevel'
        - in: query
          name: requiresReview
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Pull request list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullRequestListResponse'
  /api/v1/pull-requests/{pullRequestId}:
    get:
      tags: [PullRequests]
      summary: Get pull request detail
      operationId: getPullRequestDetail
      parameters:
        - $ref: '#/components/parameters/PullRequestId'
      responses:
        '200':
          description: Pull request detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullRequestDetailResponse'
  /api/v1/pull-requests/{pullRequestId}/risk-analysis:
    get:
      tags: [PullRequests]
      summary: Get PR risk analysis
      operationId: getPullRequestRiskAnalysis
      parameters:
        - $ref: '#/components/parameters/PullRequestId'
      responses:
        '200':
          description: Risk analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullRequestRiskAnalysisResponse'
        '202':
          description: Analysis in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncPendingResponse'
  /api/v1/pull-requests/{pullRequestId}/review-recommendations:
    get:
      tags: [PullRequests]
      summary: Get recommended reviewers and CODEOWNERS coverage
      operationId: getReviewRecommendations
      parameters:
        - $ref: '#/components/parameters/PullRequestId'
      responses:
        '200':
          description: Review recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewRecommendationsResponse'
  /api/v1/pull-requests/assist:
    post:
      tags: [PullRequests]
      summary: Prepare PR creation assistance for a source/base branch pair
      operationId: preparePullRequestAssist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PullRequestAssistRequest'
      responses:
        '200':
          description: PR assistance result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullRequestAssistResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    PullRequestId:
      in: path
      name: pullRequestId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
    RiskLevel:
      type: string
      enum: [low, medium, high, critical]
    ErrorCode:
      type: string
      description: |
        Canonical machine-readable error code. Source of truth:
        `docs/03-api/error_code_standard.md`. New codes must be added there
        and mirrored here before use.
      enum:
        - AUTH_REQUIRED
        - AUTH_TOKEN_EXPIRED
        - AUTH_TOKEN_INVALID
        - SSO_ACCESS_DENIED
        - PERMISSION_DENIED
        - REPOSITORY_ACCESS_DENIED
        - ORG_SCOPE_DENIED
        - ADMIN_REQUIRED
        - INVALID_REQUEST
        - INVALID_FILTER
        - VALIDATION_FAILED
        - IDEMPOTENCY_KEY_CONFLICT
        - GITHUB_APP_NOT_INSTALLED
        - GITHUB_APP_PERMISSION_INSUFFICIENT
        - GITHUB_API_RATE_LIMITED
        - GITHUB_API_UNAVAILABLE
        - GITHUB_WEBHOOK_SIGNATURE_INVALID
        - GITHUB_RESOURCE_STALE
        - REPOSITORY_NOT_FOUND
        - BRANCH_NOT_FOUND
        - BASE_BRANCH_REQUIRED
        - BRANCH_NAME_INVALID
        - HISTORY_GRAPH_TOO_LARGE
        - PR_NOT_FOUND
        - PR_ANALYSIS_NOT_READY
        - PR_ANALYSIS_FAILED
        - PR_BASE_BRANCH_INVALID
        - PR_REVIEWERS_REQUIRED
        - PR_CODEOWNERS_MISSING
        - PR_REQUIRED_CHECKS_FAILED
        - PR_MERGE_QUEUE_REQUIRED
        - PR_ALREADY_LINKED
        - CONFLICT_CASE_NOT_FOUND
        - CONFLICT_TYPE_UNSUPPORTED
        - CONFLICT_CONTEXT_INSUFFICIENT
        - CONFLICT_GUIDANCE_NOT_READY
        - CONFLICT_ESCALATION_REQUIRED
        - BACKOUT_NOT_FOUND
        - BACKOUT_REASON_REQUIRED
        - BACKOUT_TARGET_REQUIRED
        - BACKOUT_APPROVER_REQUIRED
        - BACKOUT_ALREADY_EXISTS
        - BACKOUT_REVERT_PR_GENERATION_FAILED
        - BACKOUT_RELEASE_POLICY_BLOCKED
        - BACKOUT_CONFLICT_PREDICTED
        - POLICY_TEMPLATE_NOT_FOUND
        - POLICY_RULE_INVALID
        - POLICY_APPLY_SCOPE_INVALID
        - POLICY_EXCEPTION_REQUIRED
        - POLICY_ENFORCEMENT_BLOCKED
        - POLICY_VERSION_CONFLICT
        - DASHBOARD_RANGE_INVALID
        - ANALYTICS_NOT_READY
        - JOB_NOT_FOUND
        - JOB_ALREADY_TERMINAL
        - JOB_TIMEOUT
        - JOB_WORKER_UNAVAILABLE
        - INTERNAL_SERVER_ERROR
        - DEPENDENCY_UNAVAILABLE
        - DATABASE_UNAVAILABLE
        - CACHE_UNAVAILABLE
        - UPSTREAM_TIMEOUT
    ApiMeta:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          description: Request tracing id, always present on API responses.
        timestamp:
          type: string
          format: date-time
          description: Server response time in ISO-8601 UTC.
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: Default user-visible message in the active locale.
            details:
              type: object
              additionalProperties: true
            retryable:
              type: boolean
              description: Whether retrying the same request is meaningful.
            userAction:
              type: string
              description: Suggested next user action, if any.
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: object
              properties:
                fieldErrors:
                  type: array
                  items:
                    type: object
                    additionalProperties: false
                    properties:
                      field:
                        type: string
                      message:
                        type: string
    UserSummary:
      type: object
      additionalProperties: false
      required: [id, login, displayName]
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
          nullable: true
...
    PullRequestSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
        title:
          type: string
        state:
          type: string
          enum: [open, closed, merged, draft]
        author:
          $ref: '#/components/schemas/UserSummary'
        baseBranch:
          type: string
        headBranch:
          type: string
        riskLevel:
          $ref: '#/components/schemas/RiskLevel'
        hasConflicts:
          type: boolean
        waitingForReview:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    PullRequestListResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PullRequestSummary'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    PullRequestDetailResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: false
          properties:
            pullRequest:
              $ref: '#/components/schemas/PullRequestSummary'
            workflowState:
              $ref: '#/components/schemas/PullRequestState'
            githubUrl:
              type: string
              format: uri
            interpretedSummary:
              type: string
            changedFiles:
              type: integer
            commits:
              type: integer
            labels:
              type: array
              items:
                type: string
            linkedBuilds:
              type: array
              items:
                $ref: '#/components/schemas/BuildSummary'
            relatedIssues:
              type: array
              items:
                type: string
            reviewStatus:
              $ref: '#/components/schemas/PullRequestReviewStatus'
            qualityGateStatus:
              $ref: '#/components/schemas/PullRequestQualityGateStatus'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    PullRequestRiskAnalysisResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: false
          properties:
            riskLevel:
              $ref: '#/components/schemas/RiskLevel'
            score:
              type: integer
              minimum: 0
              maximum: 100
            summary:
              type: string
            signals:
              type: array
              items:
                $ref: '#/components/schemas/RiskSignal'
            recommendedTests:
              type: array
              items:
                type: string
            impactedModules:
              type: array
              items:
                type: string
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ReviewRecommendationsResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: false
          properties:
            requiredCodeOwners:
              type: array
              items:
                type: string
            missingCodeOwners:
              type: array
              items:
                type: string
            recommendedReviewers:
              type: array
              items:
                $ref: '#/components/schemas/UserSummary'
            rationale:
              type: array
              items:
                type: string
        meta:
          $ref: '#/components/schemas/ApiMeta'
    PullRequestAssistRequest:
      type: object
      additionalProperties: false
      required: [repositoryId, sourceBranch, baseBranch]
      properties:
        repositoryId:
          type: string
          format: uuid
        sourceBranch:
          type: string
        baseBranch:
          type: string
        draft:
          type: boolean
          default: false
    PullRequestAssistResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: false
          properties:
            proposedTitle:
              type: string
            proposedBody:
              type: string
            recommendedReviewers:
              type: array
              items:
                $ref: '#/components/schemas/UserSummary'
            checklist:
              type: array
              items:
                type: string
            riskSummary:
              type: string
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    AsyncPendingResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: false
          properties:
            status:
              type: string
              enum: [queued, running]
            jobId:
              type: string
              format: uuid
        meta:
          $ref: '#/components/schemas/ApiMeta'
    BuildSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
        status:
          type: string
          enum: [queued, running, success, failed, cancelled]
        targetName:
          type: string
        url:
          type: string
          format: uri
    PullRequestState:
      type: string
      description: |
        Full PR business state machine. Source of truth:
        `docs/03-api/state_transition_spec.md` section 4.
      enum:
        - draft
        - open
        - under-review
        - changes-requested
        - approved
        - merge-blocked
        - ready-to-merge
        - queued-for-merge
        - merged
        - closed
        - reverted
    StaticAnalysisStatus:
      type: string
      enum: [not-required, pending, passed, failed]
    PullRequestReviewStatus:
      type: object
      additionalProperties: false
      properties:
        reviewerCount:
          type: integer
        requiredReviewerCount:
          type: integer
        approvals:
          type: integer
        changesRequested:
          type: integer
        codeOwnersSatisfied:
          type: boolean
        checklistCompleted:
          type: boolean
        missingReviewers:
          type: array
          items:
            type: string
        pendingReviewers:
          type: array
          items:
            type: string
    PullRequestQualityGateStatus:
      type: object
      additionalProperties: false
      properties:
        requiredChecksPassed:
          type: boolean
        requiredCheckCount:
          type: integer
        requiredCheckPassingCount:
          type: integer
        failingChecks:
          type: array
          items:
            type: string
        mergeBlockedReasons:
          type: array
          items:
            type: string
        staticAnalysisStatus:
          $ref: '#/components/schemas/StaticAnalysisStatus'
        commitMessagePolicyPassed:
          type: boolean
        releaseImpact:
          type: boolean
        mergeQueueRequired:
          type: boolean
...
    RiskSignal:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
        severity:
          $ref: '#/components/schemas/RiskLevel'
        summary:
          type: string
        scoreContribution:
          type: integer